+++ flux/tasks/ai/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from flux.tasks.ai.agent import agent
+
+__all__ = ["agent"]
